Fix incorrect implicit cast. This was what actually caused the pressure to go unstable, not the windspeed-pressure adjustment I removed earlier.

+++ server/modules/cfweather.cpp
@@ -686,8 +686,8 @@ static void smooth_wind() {
                 weathermap[x][y].windspeed = 0;
             }
             // The wind moves some of the higher pressure to the lower pressure.
-            weathermap[x][y].pressure -= weathermap[x][y].windspeed/(WIND_FACTOR*2);
-            weathermap[x+dx][y+dy].pressure += weathermap[x][y].windspeed/(WIND_FACTOR*2);
+            weathermap[x][y].pressure -= (int)(weathermap[x][y].windspeed/(WIND_FACTOR*2));
+            weathermap[x+dx][y+dy].pressure += (int)(weathermap[x][y].windspeed/(WIND_FACTOR*2));
         }
 
     /*  now, lets crank on the speed.  When surrounding tiles all have
